// src/Models/goals.ts
export interface Goals{
    currentWeight: number;
    goalWeight: number;
    weeklyGoal: number;
    height: number;
    activityLevel: number;
    bmr: number;
    tdee: number;
    leanBodyMass: number;
    bodyFat: number;
    goalCalories: number;
    goalFat: number;
    goalCarbs: number;
    goalProtein: number;
    // eg lose/gain/maintain
    goal: number;
    carbPercent: number;
    proteinPercent: number;
    fatPercent: number;
    percentTotal: number;
    totalMatch: any;

}